Added more casts to long long, to suppress warnings on 32-bit unix.

+++ fbmuck/src/wiz.c
@@ -990,7 +990,7 @@ do_muf_topprofs(dbref player, char *arg1)
 		free(curr);
 	}
 	snprintf(buf, sizeof(buf), "Profile Length (sec): %5lld  %%idle: %5.2f%%  Total Cycles: %5lu",
-			(current_systime-sel_prof_start_time),
+			(long long)(current_systime-sel_prof_start_time),
 			((double)(sel_prof_idle_sec+(sel_prof_idle_usec/1000000.0))*100.0)/
 			(double)((current_systime-sel_prof_start_time)+0.01),
 			sel_prof_idle_use);
@@ -1100,7 +1100,7 @@ do_mpi_topprofs(dbref player, char *arg1)
 		free(curr);
 	}
 	snprintf(buf, sizeof(buf), "Profile Length (sec): %5lld  %%idle: %5.2f%%  Total Cycles: %5lu",
-			(current_systime-sel_prof_start_time),
+			(long long)(current_systime-sel_prof_start_time),
 			(((double)sel_prof_idle_sec+(sel_prof_idle_usec/1000000.0))*100.0)/
 			(double)((current_systime-sel_prof_start_time)+0.01),
 			sel_prof_idle_use);
@@ -1272,7 +1272,7 @@ do_all_topprofs(dbref player, char *arg1)
 		free(curr);
 	}
 	snprintf(buf, sizeof(buf), "Profile Length (sec): %5lld  %%idle: %5.2f%%  Total Cycles: %5lu",
-			(current_systime-sel_prof_start_time),
+			(long long)(current_systime-sel_prof_start_time),
 			((double)(sel_prof_idle_sec+(sel_prof_idle_usec/1000000.0))*100.0)/
 			(double)((current_systime-sel_prof_start_time)+0.01),
 			sel_prof_idle_use);
